Create README.md
Data Set
For this blog, we have used the Kaggle data set — NFL Big Data Bowl.
The dataset contains 65k rows and 48 columns of NFL data. Each row in the file corresponds to a single player’s involvement in a single play. The dataset was intentionally joined (i.e. denormalized) to make the API simple. All the columns are contained in one large data frame which is grouped and provided by PlayId.